LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-change-try-rule-expected.txt


FAIL Starts rendering with --try assert_equals: expected 200 but got 100
FAIL No successful position, keep --try assert_equals: expected 250 but got 100
FAIL No successful position, last successful invalidated by @position-try change assert_equals: expected -50 but got 100
